fix: 형변환 수정

See merge request s11-webmobile1-sub2/S11P12D107!333

+++ backend/src/main/java/com/chestnut/backend/log/service/LogService.java
@@ -65,7 +65,7 @@ public class LogService {
                 .member(member)
                 .chapter(study.getChapter())
                 .study(study)
-                .todayCount((byte) (todayCount + 1))
+                .todayCount(todayCount + 1)
                 .isPass(isPass)
                 .passRecord(passRecord)
                 .build();
